requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
Markdown==3.5.1
MarkupSafe==2.1.5
openai==1.6.1
orjson==3.9.14
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.5.3
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.14.6
pyotp==2.9.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
tqdm==4.66.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
